personal-well-agent/requirements.txt
crewai
langchain
langchain-openai
openai
schedule
notify-py
twilio
python-dotenv
requests
tk